Remove mention of command line arguments
My original intent was that you could spot the use of TLS distribution because of `-proto_dist inet_tls`, but -- duh -- _we_ appear in that spot, so without introducing another command line arg (such as -proto_dist_orig or something), there's no point. If you want TLS distribution, create a copy of this module (epmd_docker_tls_dist or something), with the changed macro, and then use that instead.